backend/requirements.txt
boto3
psutil
fastapi
uvicorn
requests
watchdog
python-jose[cryptography]